requirements.txt
# Main project requirements
httpx==0.25.2
pyyaml==6.0.1
gql[requests]==3.4.1
infrahub-sdk>=0.16.0

